requirements.txt
flask==2.0.1
openai==1.3.0
speechmatics-python==2.7.0
pyaudio==0.2.13
python-dotenv==1.0.0
flask-cors==4.0.0
requests==2.31.0 